Validators extracted from SpecFlowGherkinParser private methods
