+++ TechTalk.SpecFlow.Parser/ISemanticValidator.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace TechTalk.SpecFlow.Parser
+{
+    internal interface ISemanticValidator
+    {
+        List<SemanticParserException> Validate(SpecFlowFeature feature);
+    }
+}
